통합 병합 수행_260616

+++ src/main/webapp/WEB-INF/views/party/partyboard.jsp
@@ -271,7 +271,9 @@
 
 	// AJAX 딜레이
 	let delay = 1000;
-
+	
+	let flag = false;
+	
 	$(function()
 	{
 		// 데이터 AJAX 로 바인딩
@@ -306,16 +308,30 @@
 				,"error":function(e)
 				{
 					clearInterval(interval);
-		            if (e.status === 401)
+					
+					if (e.status === 401)
 		            {
+				  		alert("로그인 안함");
 		                location.href = "${path}/user/login";
 		            }
-		            else if (e.status === 403 || e.status === 404)
+		            else if (e.status === 403)
 		            {
+		            	alert("강퇴당했습니다.");
 		                location.href = "${path}/party/list";
+		            }
+		            else if(e.status === 404)
+	            	{
+		    			alert("유효하지 않은 파티입니다.");
+		    			location.href = "${path}/party/list";
+	            	}
+		            else if(e.status == 500)
+		            {
+		            	alert("서버 오류가 발생했습니다.");
+		            	console.log(e.responseText);
 		            }
 		            else
 		            {
+		            	alert(e.responseText);
 		                console.log(e.responseText);
 		            }
 				}
@@ -585,7 +601,15 @@
 			commentInput.focus();
 			return;
 		}
-
+		
+		if(flag)
+		{
+			alert("잠시 후 다시 시도하세요");
+			return;
+		}
+		
+		flag = true;
+		
 		// alert("댓글 작성");
 		// comment.value = "";
 
@@ -612,21 +636,38 @@
 				{
 					alert("댓글 작성 실패");
 				}
+				
+				flag = false;
 			}
 			, "error":function(e)
 			{
 			  	if (e.status === 401)
 	            {
+			  		alert("로그인 안함");
 	                location.href = "${path}/user/login";
 	            }
-	            else if (e.status === 403 || e.status === 404)
+	            else if (e.status === 403)
 	            {
+	            	alert("권한이 없습니다.");
 	                location.href = "${path}/party/list";
+	            }
+	            else if(e.status === 404)
+            	{
+	    			alert("유효하지 않은 파티입니다.");
+	    			location.href = "${path}/party/list";
+            	}
+	            else if(e.stats == 500)
+	            {
+	            	alert("서버 오류가 발생했습니다. 잠시 후 다시 시도해 주세요.");
+	            	console.log(e.responseText);
 	            }
 	            else
 	            {
+	            	alert(e.responseText);
 	                console.log(e.responseText);
 	            }
+			  	
+			  	flag = false;
 			}
 		});
 	}
@@ -640,6 +681,14 @@
 			return;
 		}
 		
+		if(flag)
+		{
+			alert("잠시 후 다시 시도하세요");
+			return;
+		}
+		
+		flag = true;
+		
 		$.ajax(
 		{
 			"type":"POST"
@@ -652,23 +701,39 @@
 					alert("삭제 실패");
 					btn.disabled = false;
 				}
+				
+				flag = false;
 			}
 			, "error":function(e)
 			{
 			  	if (e.status === 401)
 	            {
+			  		alert("로그인 안함");
 	                location.href = "${path}/user/login";
 	            }
-	            else if (e.status === 403 || e.status === 404)
+	            else if (e.status === 403)
 	            {
+	            	alert("권한이 없습니다.");
 	                location.href = "${path}/party/list";
 	            }
+	            else if(e.status == 404)
+	            {
+	            	alert("유효하지 않은 파티입니다.");
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 500)
+			  	{
+	            	alert("서버 오류가 발생했습니다. 잠시 후 다시 시도해주세요.");
+	            	console.log(e.responseText);
+			  	}
 	            else
 	            {
-	            	alert("서버 오류");
+	            	alert(e.responseText);
 	                console.log(e.responseText);
-	                btn.disabled = false;
 	            }
+			  	
+				btn.disabled = false;
+			  	flag = false;
 			}
 		});
 	}
@@ -677,6 +742,14 @@
 	{
 		//alert("레디");
 		// ajax 레디 요청
+		
+		if(flag)
+		{
+			alert("잠시 후 다시 시도하세요");
+			return;
+		}
+		
+		flag = true;
 		
 		btn.disabled = true;
 		
@@ -692,24 +765,38 @@
 					alert("레디 실패");	
 				}
 				
+				flag = false;
 				btn.disabled = false;
 			}
 			,"error":function(e)
 			{
 				if (e.status === 401)
 	            {
+					alert("로그인 안함");
 	                location.href = "${path}/user/login";
 	            }
-	            else if (e.status === 403 || e.status === 404)
+	            else if (e.status === 403)
 	            {
+	            	alert("권한이 없습니다.")
 	                location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 404)
+            	{
+	        		alert("유효하지 않은 파티입니다.");
+	        		location.href = "${path}/party/list";
+            	}
+	            else if(e.status == 500)
+	            {
+	            	alert("서버 오류가 발생했습니다. 잠시 후 다시 시도해 주세요.");
+	            	console.log(e.responseText);
 	            }
 	            else
 	            {
-	            	alert("서버 오류");
+	            	alert(e.responseText);
 	                console.log(e.responseText);
-	                btn.disabled = false;
 	            }
+				btn.disabled = false;
+                flag = false;
 			}
 		});
 		
@@ -722,6 +809,14 @@
 			btn.disabled = false;
 			return;
 		}
+		
+		if(flag)
+		{
+			alert("잠시 후 다시 시도하세요");
+			return;
+		}
+		
+		flag = true;
 		
 		//alert("승인");
 		// ajax 승인 요청
@@ -737,24 +832,38 @@
 				{
 					alert("승인 실패");
 					btn.disabled = false;
-				}				
+				}			
+				flag = false;
 			}
 			, "error":function(e)
 			{
 				if (e.status === 401)
 	            {
+					alert("로그인 안함");
 	                location.href = "${path}/user/login";
 	            }
-	            else if (e.status === 403 || e.status === 404)
+	            else if (e.status === 403)
+	            { 
+	            	alert("권한이 없습니다.")
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 404)
 	            {
-	                location.href = "${path}/party/list";
+	            	alert("유효하지 않은 파티입니다.");
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 500)
+	            {
+	            	alert("서버 오류가 발생했습니다. 잠시 후 다시 시도해 주세요");
+	            	console.log(e.responseText);
 	            }
 	            else
 	            {
-	            	alert("서버 오류");
+	 				alert(e.responseText);
 	                console.log(e.responseText);
-	                btn.disabled = false;
 	            }
+				btn.disabled = false;
+                flag = false;
 			}
 		});
 	}
@@ -767,6 +876,13 @@
 			return;
 		}
 		
+		if(flag)
+		{
+			alert("잠시 후 다시 시도하세요");
+			return;
+		}
+		flag = true;
+		
 		$.ajax(
 		{
 			"type":"POST"
@@ -778,32 +894,51 @@
 				{
 					alert("거절 실패");
 					btn.disabled = false;
-				}				
+				}		
+				flag = false;
 			}
 			, "error":function(e)
 			{
 				if (e.status === 401)
 	            {
+					alert("로그인 안함");
 	                location.href = "${path}/user/login";
 	            }
-	            else if (e.status === 403 || e.status === 404)
+	            else if (e.status === 403)
+	            { 
+	            	alert("권한이 없습니다.")
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 404)
 	            {
-	                location.href = "${path}/party/list";
+	            	alert("유효하지 않은 파티입니다.");
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 500)
+	            {
+	            	alert("서버 오류가 발생했습니다. 잠시 후 다시 시도해 주세요");
+	            	console.log(e.responseText);
 	            }
 	            else
 	            {
-	            	alert("서버 오류");
+	 				alert(e.responseText);
 	                console.log(e.responseText);
-	                btn.disabled = false;
 	            }
+				btn.disabled = false;
+                flag = false;
 			}
 		});
 	}
 
 	function reservation()
 	{
-		alert("예약");
+		// alert("예약");
 		// 예약 페이지 이동
+		
+		if(confirm("예약하시겠습니까?"))
+		{
+			window.location.href = "${path}/party/reservation/" + partyId;
+		}
 	}
 
 	function partyUpdate()
@@ -821,6 +956,12 @@
 		
 		if(confirm("파티를 해산 하시겠습니까?"))
 		{
+			if(flag)
+			{
+				alert("잠시 후 다시 시도하세요");
+				return;
+			}
+			
 			window.location.href="${path}/party/delete/" + partyId;
 		}
 	}
@@ -833,6 +974,14 @@
 			return;
 		}
 		
+		if(flag)
+		{
+			alert("잠시 후 다시 시도하세요");
+			return;
+		}
+		
+		flag = true;
+		
 		$.ajax(
 		{
 			"type":"POST"
@@ -844,24 +993,38 @@
 				{
 					alert("강퇴 실패");
 					btn.disabled = false;
-				}				
+				}			
+				flag = false;
 			}
 			, "error":function(e)
 			{
 				if (e.status === 401)
 	            {
+					alert("로그인 안함");
 	                location.href = "${path}/user/login";
 	            }
-	            else if (e.status === 403 || e.status === 404)
+	            else if (e.status === 403)
+	            { 
+	            	alert("권한이 없습니다.")
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 404)
 	            {
-	                location.href = "${path}/party/list";
+	            	alert("유효하지 않은 파티입니다.");
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 500)
+	            {
+	            	alert("서버 오류가 발생했습니다. 잠시 후 다시 시도해 주세요");
+	            	console.log(e.responseText);
 	            }
 	            else
 	            {
-	            	alert("서버 오류");
+	 				alert(e.responseText);
 	                console.log(e.responseText);
-	                btn.disabled = false;
 	            }
+				btn.disabled = false;
+                flag = false;
 			}
 		});
 	}
@@ -874,6 +1037,13 @@
 			return;
 		}
 		
+		if(flag)
+		{
+			alert("잠시 후 다시 시도하세요");
+			return;
+		}
+		flag = true;
+		
 		$.ajax(
 		{
 			"type":"POST"
@@ -885,6 +1055,7 @@
 				{
 					btn.disabled = false;
 					alert("탈퇴 실패");
+					flag = fasle;
 				}
 				else
 				{
@@ -895,18 +1066,31 @@
 			{
 				if (e.status === 401)
 	            {
+					alert("로그인 안함");
 	                location.href = "${path}/user/login";
 	            }
-	            else if (e.status === 403 || e.status === 404)
+	            else if (e.status === 403)
+	            { 
+	            	alert("권한이 없습니다.")
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 404)
 	            {
-	                location.href = "${path}/party/list";
+	            	alert("유효하지 않은 파티입니다.");
+	            	location.href = "${path}/party/list";
+	            }
+	            else if(e.status == 500)
+	            {
+	            	alert("서버 오류가 발생했습니다. 잠시 후 다시 시도해 주세요");
+	            	console.log(e.responseText);
 	            }
 	            else
 	            {
-	            	alert("서버 오류");
+	 				alert(e.responseText);
 	                console.log(e.responseText);
-	                btn.disabled = false;
 	            }
+				btn.disabled = false;
+                flag = false;
 			}
 		});
 	}
